packages/robot-up/pyproject.toml
[project]
name = "petriplate-robot-sim"
version = "0.1.0"
description = "Headless Mujoco simulation broadcaster over websocket"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "mujoco>=3.2.6",
  "pillow>=11.0.0",
  "websockets>=12.0",
]
